+++ ghcide.cabal
@@ -1,4 +1,4 @@
-cabal-version:      1.20
+cabal-version:      3.0
 build-type:         Simple
 category:           Development
 name:               ghcide
@@ -175,6 +175,41 @@ library
         Development.IDE.GHC.HieBin
     ghc-options: -Wall -Wno-name-shadowing
 
+library ghcide-session-loader
+    visibility: public
+    default-language: Haskell2010
+    hs-source-dirs: session-loader
+    exposed-modules:
+        Development.IDE.Session
+    ghc-options: -Wall -Wno-name-shadowing
+    build-depends:
+        base == 4.*,
+        aeson,
+        async,
+        base16-bytestring >=0.1.1 && <0.2,
+        bytestring,
+        containers,
+        cryptohash-sha1 >=0.11.100 && <0.12,
+        directory,
+        extra,
+        filepath,
+        ghc,
+        ghc-check,
+        ghc-paths,
+        ghcide,
+        haskell-lsp,
+        haskell-lsp-types,
+        hie-bios >= 0.6.0 && < 0.7,
+        safe-exceptions,
+        shake,
+        text,
+        time,
+        unordered-containers
+    -- We can't build this with ghc-lib since it interfaces with hie-bios which
+    -- uses plain ghc types
+    if flag(ghc-lib)
+        buildable: False
+
 executable ghcide-test-preprocessor
     default-language: Haskell2010
     hs-source-dirs: test/preprocessor
@@ -237,24 +272,12 @@ executable ghcide
                 "-with-rtsopts=-I0 -qg -A128M"
     main-is: Main.hs
     build-depends:
-        time,
-        async,
-        hslogger,
-        aeson,
         base == 4.*,
-        binary,
-        base16-bytestring >=0.1.1 && <0.2,
-        bytestring,
-        containers,
-        cryptohash-sha1 >=0.11.100 && <0.12,
         data-default,
-        deepseq,
         directory,
         extra,
         filepath,
-        ghc-check >= 0.5.0.1 && < 0.6,
-        ghc-paths,
-        ghc,
+        ghcide-session-loader,
         gitrev,
         hashable,
         haskell-lsp,
@@ -262,8 +285,6 @@ executable ghcide
         hie-bios >= 0.6.0 && < 0.7,
         ghcide,
         optparse-applicative,
-        safe-exceptions,
-        shake,
         text,
         unordered-containers
     other-modules:
